requirements.txt
# Railway Deployment Requirements - Complete Backend Dependencies
streamlit>=1.28.0
pandas>=1.5.0,<2.1.0
numpy>=1.24.0,<2.0.0
pandas-ta>=0.3.14b0
plotly>=5.15.0

# Core Python packages
setuptools>=68.0.0,<81.0.0
wheel>=0.41.0
pip>=23.0.0

# API clients
requests>=2.31.0
aiohttp>=3.8.0

# Trading APIs
alpaca-trade-api>=3.0.0
alpaca-py>=0.8.0
polygon-api-client>=1.12.0
yfinance>=0.2.18

# AI/ML libraries
openai>=1.0.0
anthropic>=0.3.0
scikit-learn>=1.3.0
numba>=0.57.0

# Utilities
python-dotenv>=1.0.0
schedule>=1.2.0
pytz>=2023.3

# Logging and monitoring
loguru>=0.7.0

# Web framework for health checks and webhooks
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
websockets>=9.0,<11.0

# Database (MongoDB)
pymongo[srv]>=4.0.0
motor>=3.3.0

# Enhanced API features
python-multipart>=0.0.6

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0

# Production monitoring
psutil>=5.9.0
websocket-client>=1.6.0

# Additional devcontainer dependencies
ipython>=8.0.0
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Production reliability packages (optional but recommended)
praw>=7.7.0  # Reddit API for sentiment analysis (graceful degradation if not available)
textblob>=0.17.1  # Text processing for sentiment analysis (graceful degradation)
tweepy>=4.14.0  # Twitter API (graceful degradation if not available)
